backend/requirements.txt
# Backend core
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
httpx==0.26.0

# Data Science
scikit-learn==1.1.3
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
joblib==1.3.2

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# File handling
python-multipart==0.0.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4 